Enemy Flashing / Damage Update

+++ Interfaces/IEnemy.cs
@@ -7,7 +7,7 @@ namespace LegendOfZelda
     public interface IEnemy: IUpdateable, ICollidable
     {
         void Spawn();
-        void UpdateHealth(int damagePoints);
+        void UpdateHealth(float damagePoints);
         void Attack();
         void ChangePosition();
         void ChangeDirection();
